restructured test packages to match production packages
