+++ src/test/java/guitests/TagEditDialogTest.java
@@ -1,6 +1,7 @@
-package address.guitests;
+package guitests;
 
 import javafx.scene.input.KeyCode;
+
 import org.junit.Test;
 
 import static org.testfx.api.FxAssert.verifyThat;
